Change - additional development of picaps_test_node
This largely brings in the guts of the old kickstart script used to
build the original PICAPS servers and wraps it for puppet execution in
a quick-n-dirty fashion.

This is a horrible hack, but acceptable so long as these servers are
short-lived as promised.  There is very little here that is idempotent,
so the kickstart portion needs to work correctly the first time and then
not be run again.