Fix - disabling selinux breaks well known symlink
Ordinarily, Fedora has it such that /etc/sysconfig/selinux is a symlink
to /etc/selinux/config and the former is what most Fedora admins expect
to reference.  Our usage of sed broke that link and left in its place
the correct looking version, but the unused one as SELinux only honors
the latter.

- reference canonical config instead of conventional
- tell sed to use copy instead of rename when shuffling files