Change - lighten load on puppetmaster
The puppetmaster has lately been mailing many errors like:

    Could not retrieve catalog from remote server: Error 400 on SERVER ...

This is apparently due to the single-threaded nature of the embedded
Webrick http server.  The best practice for scaling is to use
Passenger/Apache instead, which is sort of like a mod_ruby.  It is not
yet possible to install Passenger on Fedora (using only std. yum
repositories) due to upstream licensing issues that are being worked
out.

In the meantime, this alleviates the problem by reducing the run
interval of clients from twice an hour to just once an hour.