New - bacula client: mdct-aos-master-f15
While the host can be rebuilt entirely from a fresh Fedora install and
the puppet manifest, the host still has the entire AOS conf.d pool to be
concerned about.  Even that has a bare git repository on mdct-00fs that
also gets bacula love, but if the "git push" via commit-resources (of
the mdct-puppeteer-admin package) is failing for some reason, this
backup would be the last line of defense.  So yes, a good thing.