Fix - Fedora 16 requires pam_krb5
While this package has been available prior to F16, it was not required
for MDC's authentication against Dart's AD servers.  For yet unknown
reasons, it is now required as of F16.  This is being applied
unconditionally since it seems harmless on older releases and keeps the
dependencies more simply stated.