Fix - Kerberos authentication occasionally fails
When krb5_server is set, sssd will contact only the one server to which
DNS resolves.  Some of our AD domain controllers reside within the DMZ
and if DNS resolves to one of these, authentication will be impossible
as those servers will always be unreachable.  Likewise, the round-robin
nature of DNS may resolve to an AD DC that is temporarily unreachable
due to maintenance, WAN outage, etc..

This revision leaves krb5_server undefined, which apparently instructs
sssd to use the realm name and more importantly perform its own
automatic fail-over from one server to the next when failures are
encountered.