Revert "Fix - puppet falsely claims postgres user non-existent"
This reverts commit 597a58ff11c45d45f7a7de6bf3922df6d7bbac27.

We are in the process of deprecating the LDAP-based postgres user and
group accounts.  Local user and groups have been created on all known
affected hosts so that the user/group lookup will now succeed.

NB: The actual lookup failure was due to the min_id setting for sssd,
not with puppet or timeouts as originally suspected.  (The LDAP account
was numbered 40/233 and min_id was set to 500.)