New - node mdct-00sw
This node acts as a Shared Workstation which might be exploited in
a variety of ways.  Presently, the use case is for PICAPS functional
testing where various tools such as test-automation might be used by
employees who don't have (or want) a Linux system of their own from
which to perform these activities.

While this establishes only one user, more can certainly be added at any
later time if needed.