Change - sudo password policy
Because the MDCT is generally very good about locking terminal sessions
with password-protected screen savers and the like, the default
requirement of having sudo require a password to gain a ticket is
a little abrasive.  Levi Harper suggested we drop this requirement to
make it's use more acceptable.

This revision also drops the custom ticket time-out setting as it is
mostly irrelevant in this new configuration.