Change - update bacula::client for systemd
Recent updates to the bacula-client package now have USER and GROUP
settings in /etc/sysconfig/bacula-fd that were not there previously.
This updates the puppet resource for that file to ensure that those are
set appropriately.  Without this the client likely will not start
correctly, as was definitely observed on mdct-dev12 today (despite being
systemd-based now since Fedora 15, oddly).