Change - kdebase renamed to kde-baseapps
Midway into the F16 life cycle, this package got renamed.  This is not
an ideal situation because the dependency is really on the KDE version,
not the Fedora version.  However, we don't have a KDE version fact so
this will have to do.  Provided F16 systems are updated, this should
suffice.