Change - add debugs to aid problem isolation
As of Fedora 16, the install-and-setup script is having issues with the
configuration of MySQL.  It is hoped that these debugging messages might
assist in isolating the exact cause during the next builds.

Here's what's known at this time:
    - the mysql-server class successfully installs and starts the server
    - the setup-and-install script then fails to execute the service
      unit's ExecStartPre=/usr/libexec/mysqld-prepare-db-dir for unknown
      reasons
    - manual confirmation of those exact same steps after the failure
      seem to work correctly

One theory is that the TimeoutSec feature of the service unit is not
being honored by systemd and the install-and-setup script is attempting
to relocate the data dir and restart the service before the original
instance (started by mysql-server class) has had sufficient time to
terminate.