New - puppet-2.7 compatibility tool
This tool has been developed to automatically adjust all puppet
manifests in use at Dart to accommodate a transition to
puppet-server-2.7.13-1.fc17.noarch (from
puppet-server-2.6.16-1.fc15.noarch).  An unintentional revision early in
the 2.7 series broke the 'import' statement that had once been required,
but which is now apparently frowned upon.  Allegedly, it was not the
intention to deprecate this usage, but there appears to be little effort
being made to actually resolve the issue.  Thus this tool was developed
to make Dart's puppet modules compliant with that shipped in Fedora 17
(shown above).

For more details, see http://projects.puppetlabs.com/issues/8433.