New - selinux module
This merely provides an abstraction layer around puppet's native features
that tend to oscillate between working and not working, depending on the
component versions, tide levels and the whim of evil spirits.  By using
this layer instead of the native offerings, adapations and kludges can
target the entire codebase by doing so here.

This module provided by jflorian@doubledog.org under the terms of the
GNU Public License, version 2.