Fix - dependency may be unenforceable
If selinux is disabled on a host, the boolean will not be managed or
even declared.  If it's not declared, it can be used as a dependency.
This revision reverses the dependency reference so that if it the
boolean is set, it must be done before the service is managed, which is
effectively the same but with the potential for the error.