Change - label for IDEA release extraction exec
This removes the 'idea' suffix from the exec of tar and replaces it with
the build number instead.  The former was redundant given the $name
passed for the release and the latter might help an admin recognize
a bone-headed mistake of changing the build parameter to 'absent'
instead of assigning that to the ensure parameter.  ;-)