New - mirrmaid to mirror PICAPS data for mdct-95pi and mdct-98pi
A decision has been made to consider data on hosts mdct-95pi and
mdct-98pi vital and in need of inclusion in our Bacula backups.  While
these hosts are local making a direct bacula installation feasible, this
approach maintains consistency with our other PICAPS servers (which are
not all local).