New - mirrmaid::config definition
This new definition is now the preferred way to install a mirrmaid
configuration file.  Host specific configs have been removed from the
main class and refactored as definition usages within the appropriate
host class.

The cron job was evaluated for the same treatment, but since the one job
file can easily have multiple mirrmaid invocations, it was left as is.