Fix - module facts not deployed
This revision is the result of a fairly thorough review of custom fact
handling, which was triggered by looking into our selinux module's
selinux::boolean definition which was to operate conditionally due to
a bug in puppet (http://projects.puppetlabs.com/issues/9054).  However,
a condition based on the $selinux fact would not work on many releases
of Fedora due to bugs (https://projects.puppetlabs.com/issues/15049 and
https://projects.puppetlabs.com/issues/11531).  So, I developed a custom
fact named selinux_simple as part of the selinux module, but this could
not be deployed to older versions of puppet due to more bugs, this time
with pluginsync.

The general gist here is to enable the pluginsync feature, but only if
the puppet version is something greater than 2.6.6, since 2.6.6 and
prior all had major bugs.