Fix - log ownership can become corrupted
If root mistakenly tries to run this script -- which is not supported
and won't work -- the log file could become owned by root making it
impossible for the log to work correctly thereafter when the tool was
used by the intended user.  This revision incorporates the user's name
into the log file name so that such mishaps don't require intervention
to resolve the issue.