Fix - tool generates many errors falsely
Using the --archive option meant that this tool would attempt to make
root the owner of the targets, just like the source.  Due to the way the
source and target volumes are mounted, this will never work any file
requiring sync'ing will generate misleading errors.  This revision keeps
those features of --archive which are applicable, but drops all others
to eliminate all such messages.