Change - provide error stream separately
The stderr output of teamcenter-sync is no longer combined with stdout.
Instead it is redirected to a separate log file whose contents are
incorporated directly into the body of the resultant mail so that they
may be easily reviewed without necessitating a search for them within
the primary activity log which can be very long.