New - incorporate rsync stats within mail body
The statistics that rsync dumps at the end of the main activity log are
now included also in the main mail body along with the other quick
summary details making this useful info more accessible to the reader.

Combined with the prior commit, it should be unnecessary for the
reviewer to examine the main activity log attachment in most cases.