Fix - rsync_options not seen as options by rsync
A "set -x" just before the rsync issuance revealed that this variables
contents were surrounded by single quotes which prevented rsync from
seeing --dry-run as the option it was intended to be.  Instead rsync saw
this as an argument, the first of two sources which led to some unusual
looking error messages.