New - mirroring of Fedora releases/updates via mirrmaid
This is intended to replace our motley collection of rsync scripts for
mirror management.  Specifically, this commit obsoletes those scripts
responsible for all aspects of maintaining the MDCT mirrors of Fedora
YUM repositories and covers both the base release and updates (stable
and testing) for the following Fedora branches: 17 and 16.  This
supplements existing support for Fedora development branches.