Fix - no mirrmaid config for F18 updates/testing
During the development phase, there are lots of updates in testing that
would be good to have mirrored locally so that we may do our own trials,
if needed.  These were not mirrored prior to this revision.  (We had
a batch, but they were not getting maintained.)

When updates in testing earn enough karma, they move into the base
(development/18) rather than stable updates (updates/18) because 18 is
still in the development phase.  This mirror (stable updates) was not
getting maintained either, although it matters little since the
repository is empty.  However, this config should be in place so that
when F18 is released, this mirror is ready to go.