Drop - ConnectTimeout setting for herd ssh
This option was originally added to make possible batch operations
across large portions of the flock.  Since an offline node could delay
the batch operation excessively, this option addition had merit.

With herd gaining the -a/-A options there's little need to for this
option now.  It's simply better to rely on TCP timeouts instead.