Change - teamcenter access controls
The teamcenter-sync tool was issuing numerous (thousands to tens of
thousands) of error messages pertaining to the inability to set time
stamps or permissions on both real target objects and temporary file
system objects created during run-time.  In an attempt to resolve these
issues, this revision attempts to defeat any semblance of security
enforcement over these objects from the client's (tc-util) perspective,
instead delegating all such control to the Windows servers themselves.

This works great for mas-cad16, but the forced dir_mode and file_mode
seem to have no bearing on mas-cad26 and mas-cad27.  It is my belief
that the differences rest with the configuration of those hosts and must
be address there to resolve any remaining issues.