Change - mirrmaid upstream source for Fedora
It appears that Duke is lagging a little behind with their updates, so
we're back to kernel.org for now.

IIRC, kernel.org wasn't carrying the Fedora 18 development branch at one
time, but Duke was.  If this becomes an issue, it may be best to use
distinct upstream sources, depending on branch.