Refactor - xorg-server now parameter driven
This class has been revised such that it now accepts two parameters:

    config  => Source URI to xorg.conf to be used
    drivers => List of driver packages required

Both are optional and default to doing nothing.  This module is no
longer included in the abstract base_node class.  Instead it must be
included directly on the particular host(s) affected, which has been
preserved for those already provisioned via this class.