Change - rewrite autofs module
Module now uses a params.pp manifest to isolate all the OS variant
specifics.  Notification events and requirements are now of a much more
robust design acquired from years of experience.  Host-private files
have been relocated so as to not be part of this module, but instead of
the general purpose file-server namespace.