New - autofs::mount definition
This new definition allows for concise declaration of automount
configuration files without altering the defaults for the File resource,
which in turn resolves an issue of trying to apply the wrong seltype to
the /pub symlink (which expects either 'default_t' or 'etc_runtime_t',
depending on some condition that I couldn't hunt down in a reasonable
amount of time.  The default context usually seems appropriate so this
in effect causes puppet to just ignore that aspect.