Change - kde-plasma-networkmanagement package
This package is now installed via the new network module, but only if
NetworkManager is being used.  It needs to be removed here because if
NetworkManager is not being used, a loop is established with this
installing it and the network module removing it.  Puppet cannot see the
circular loop itself because of the yum::remove hack (which necessarily
works around a puppet deficiency.)