Change - Florian's git config
Now using ~ in excludesfile instead of canonical path so that the same
file content can be used for myself and as root.

Push now configured to use 'current' method.  It would be preferable to
use 'simple' instead, but we still have too many old Fedora's around
that don't offer this.