New - manage journald.conf
Fedora 15 introduced systemd and Fedora 17 introduced the journal to
systemd.  In Fedora 18, the journal became non-persistent by default.
Dart users need/expect journaled log messages to persist through
reboots, so the default configuration for F18 and later is now revised
to force a persistent journal.

It does not seem possible (and certainly seems inadvisable) to restart
the systemd service due to such a change, so no such attempt is made
here.