Fix - network class forces legacy service with NetworkManager
This revision makes the Service resource dynamic depending upon the
value of the network_manager parameter.  This probably does not handle
well the situation of transitioning from the legacy network service to
the NetworkManager service.  (I don't have time to make it that perfect
right now and it would be very rare for this to ever occur.)