Change - set password expiration warning
It is not clear if Fedora's default changed or someone reconfigured our
Active Directory servers, but suddenly everyone our team is seeing
warnings that they're password is going to expire in NN days, even as
much as 75 days.  This client side configuration is now explicitly set
at 14 days which seems reasonable.