Change - simplify rpmfusion mirroring
This revision replaces the explicit list of rpmfusion mirrors to be kept
with an implicit one asking for all of "free" and "nonfree" for Fedora,
but then explicitly excludes the older releases we either don't want or
don't want managed by mirrmaid.  These changes were tried with the
dry-run option initially to confirm there was no hidden surprises.  The
original goal was to get the setup rpms via the convenient symlinks kept
directly in 'releases/' to better match up-stream's instructions for
proper configuration.