Change - simplify yum setup for RPM Fusion
With recent changes to the mirrmaid strategy for mirroring RPM Fusion's
repositories, we now have (local, mirrored) access to the recommended
rpms to be used for configuring the RPM Fusion repositories.  This
revision accomplishes that, but must also drop support for Fedora 13
(WRT to RPM Fusion only) since they don't provide the required content.