Change - tc-util mount credentials
A new account and password have been established for which tc-util is to
employ for mounting the various Teamcenter volumes.  This account has
been set up exclusively for this use so that when the tcadmin account's
password needs to be changed, it won't adversely affect the goals of
tc-util.

This new 'tcsync' account was actually established as of 2013-04-02 by
Chris Kent (see email dated same @ 14:02), but for some inexplicable
reason, hadn't gotten applied here until now.