Fix - virtualization support for Fedora 19
Running virt-manager on a Fedora 19 might prompt the user to install the
libvirt-daemon-kvm package.  Apparently this is no longer pulled in as
a dependency so it is now explicitly stated as such here.  The package
first became available in Fedora 17, hence the condition as written as
such, even though Fedora 17 and 18 seemed untroubled with this
requirement.