Fix - Puppet/Fedora issues with seltype on links
Older versions of Fedora and/or Puppet have difficulty managing SELinux
types on symbolic links for unclear reasons.  An endless stream of
tagmail reports can be generated as the type bounces back and forth.  It
seems to work just fine with what's available in Fedora 19 and at
present, the only known case where seltype on /pub is critical is
a Fedora 19 host.  It probably works well on other older releases
(likely F16-18), but that would require further testing that isn't
warranted now and will find diminishing value over time anyway.