Change - abstract Dart-specifics out of yum module
Everything specific to Dart has been moved out of the yum module making
it site/domain agnostic.  Specifically, classes have been renamed as
follows:

    yum::fedora     ==>     dart::subsys::yum::fedora
    yum::mdct       ==>     dart::subsys::yum::mdct
    yum::rpmfusion  ==>     dart::subsys::yum::rpmfusion

The main yum.conf file is no longer assumed to be in one of several
locations, instead the yum class now expects its URI to be a passed via
a class parameter.

Lastly, this change also splits off the RPM Fusion configuration so that
it now only applies to subclasses of dart::abstract::workstation_node
(or any other that might needed it in the future).