New - oracle::jdk support of ensure => latest
To help prevent rampant proliferation of specific oracle::jdk
version/updates littered throughout the puppet manifests, this
definition now supports 'latest' for the ensure parameter.

Note that until Oracle provides a proper yum repository (i.e., when hell
freezes, thaws and then refreezes) this will not be magical.  The
concept of 'latest' is only as fresh as this definition might be kept
updated.