New - systemd::unit $restart_events parameter
This new parameter makes it possible for a systemd::unit definition to
effectively subscribe to one or more resources for event notifications
for the purposes of restarting the unit when those events do occur.

NB.  Puppet's 'subscribe' parameter is a meta-parameter that can be
applied to the entire systemd::unit definition, but that causes all
execs here to be reapplied which is generally not the desired outcome.
This work-around seems to function quite nicely though.