New - enforce max-deletion in mirrmaid configs
rsync 3.0.0 introduced a --max-delete option to prevent deletion of more
than a specified number of files or directories.  If the limit is
exceeded, a warning message is issued and rsync exits.  This option is
now being included for all our mirrmaid configurations to prevent
catastrophic losses for any number of reasons.  The most common
situation seen to date is where the Fedora Project archives a release
after it has reached end-of-life, but before we get around to disabling
our mirroring of it.

I've not tested this change at all to learn if rsync deletes the
specified number of files and then exits or if it exits before deleting
any files if it sees that it would have to delete beyond the limit.
I suspect it's the latter that was implemented, but in any case, this
option can be seen as improvement to our existing setup.