Drop - packages::kde class
This has never been a good way to ensure that KDE Plasma Workspaces is
installed on a workstation due to the ever shifting package names and
the likelihood that this list of packages is woefully out of date.
A better alternative would be to somehow achieve the idempotent
equivalent of:

    yum groupinstall "KDE Plasma Workspaces"

Until then, this is just being dropped.  It is up to the administrator
to ensure that the above group is installed, either through anaconda
(the Fedora installer) or via the above command.