Fix - TeamCity cannot run mock successfully
When running mock as a non-root user (the only sane way), it is
imperative that the sbin paths appear after the bin paths in PATH
otherwise a mock exits with a error message stating this fact.
Unfortunately, tito hid this message (when doing 'tito release ...'),
but alas it's known and corrected now with this revision.