Change - move all koji work files to NFS
For the initial development of the Koji setup it was simpler to avoid
having these resources on NFS and the additional complexity that would
bring -- Koji was difficult enough already.  With the basic setup in
place and working, it's time to move these to NFS where it would be
possible to add more build hosts and proceed towards having mash
generate/maintain our yum repositories.