Fix - Koji can be started in wrong sequence
During the migration of the koji file resources from the local file
system to the NFS server (via rsync), I had the kojira, kojid, and httpd
services all quiesced.  Upon running puppet to effect the configuration
changes and restart the services I found that kojira failed to start
because the hub was unreachable.  This resolves that dependency and
several other related cases.