New - autofs::mount_point and autofs::map_entry
This is practically a complete overhaul of the autofs module.  The main
autofs class now accepts a 'legacy' parameter which defaults to true for
maintaining backwards compatibility (until all old usages can be
migrated).  The new mount_point and map_entry definitions offer much
improved flexibility.  This also drops the nascent autofs::map2 which
was supposed to accomplish the same thing, but fell short in advanced
testing.