Refactor - replace rpcidmapd module with nfs module
This is taking place to allow for the introduction of a new rpcgssd
class that also will require the nfs-utils package.  Rather than forcing
all that within the extant class (and making it poorly named), this
route was chosen to keep distinct classes, yet allow better sharing of
resources without violating puppet's rules.  It will also be good to
gather all NFS related resources into a single module.