Fix - firewalld removal can fail
After installation, firewalld-config has a dependency upon firewalld so
yum must be used for the removal.  (Specifying firewalld-config in
conflicting_packages doesn't work well because it didn't exist in older
Fedora's.  In fact, it is not even possible to install in the normal
fashion.  It is somehow magically installed via comps as part of the
base image.)