Fix - HTTP caching issue with yum downloads
Thanks go to Levi Harper for this find.  It specifically helps with
a situation where mdct-proxy was doing the wrong thing in handling his
oVirt repositories that are accessed on the Internet.  It also likely
will help resolve a rarer case I've seen when doing repetitive RPM
builds using the same ENVR during application package development.