New - allow_clients parameter for ntp class
This new parameter permits variable configuration of the chrony service
with respect to which clients, if any, may use the configured host as
a time source via NTP.  Presently, this parameter is only supported if
the host uses chronyd vs. the legacy ntpd since our use case only
requires that much.