New - mirrmaid support for Fedora 20
- added Fedora 20
- dropped Fedora 18
- exclude ARM architecture(s)

Technically, F18 hasn't reached EoL yet, but will in another 30 days or
so.  It's being removed from mirrmaid so we don't risk accidental
unintended loss when Fedora moves 18 from the regular area to the
archive area.  If anyone really wants a more up-to-date F18, they can
always sync it manually.