Drop - max-delete option in mirrmaid configs
This effectively reverts 57930d6a as it proves mostly useless to us.
Empirical evidence reveals that rsync deletes at most N files and then
exits.  Given how often mirrmaid runs, we can still lose an entire
Fedora release over a weekend, if not overnight.  It would be great to
revisit this concept if ever rsync introduces a control that would
prevent any deletions if the count were to exceed N.