Change - puppet preferred serialization format
For a while it was necessary to configure puppet clients so that they
informed the master that their preferred serialization format was YAML
because the F19-based master was incapable of recognizing that some
clients could do nothing else.  (F19 brought puppet-3 which started
using PSON by default, instead of YAML.)  Apparently upgrades to the
puppet master (on F19) made setting the preferred_serialization_format
unnecessary, but still harmless.  Now with the puppet master on F20,
this setting is actually problematic as clients will fail to apply the
catalog, hence the option is being dropped here completely.

This change was tested on Fedora clients ranging from release 14 through
20 and found to be happy for all.  All these clients were puppet-2.6 or
newer.  Fedora 13 has the ancient 0.25 version and other modules already
prevent puppet use on those so only Fedora 14 and newer are going to
receive any love here henceforth.