Change - configure SMTP server for puppet master
Fedora 20 no longer installs an MTA (e.g., sendmail) by default, thus
tagmail was not being sent by the puppet master.  This is a desirable
feature, but we don't necessarily need to install sendmail just for
this.  Instead we simply tell the puppet master of our MX, which can be
easily found with "dig dart.biz MX" as was done here.