Fix - PyCharm RC naming collisions
JetBrains names the tarballs distinctly for release candidates, but they
have a default extraction point that reflects the eventual release
number which can result  in collisions.  This means we couldn't have
multiple RCs installed (for the same proposed release).

This patch resolves that issue by ensuring that the extraction point
always reflects the build number.