Fix - Could not find command 'newaliases'
With the introduction of Fedora 20 where sendmail (or any MTA) is no
longer part of the default install, it is now possible that mail aliases
might be configured before sendmail is installed, thereby making it
impossible to run the newaliases command.

This resolves that issue by replacing the site default for the mailalias
resource type with an explicit, formal sendmail module to install the
package, configure the service and provide a sendmail::alias definition
to wrap the standard mailalias resource type.  This definition takes the
same parameters as the standard type, but also handles the execution of
the newaliases command to refresh the compiled alias database.

All extant use cases are hereby modified to comply.