Fix - sendmail::alias may trigger duplicate Exec
The Exec of $newaliases_cmd needs to occur for each sendmail::alias
declared, but the exec itself need only be declared once.  In fact, it
caused errors with more than one alias.  The alternative resolution here
would be to add $name to the Exec's name, but that just creates a larger
resource tree unnecessarily.