Change - migration to new rsync::server - part 1
This effects the migration of all instances of the
plant_utility_server_node class from the obsoleted rsync_server to the
new and much improved rsync::server.

This has made it possible to collapse the numerous and repetitious
private-host files into a pair of shared files under the dart module.