New - firewall-driven support
This establishes a new firewall-driven module for puppet to manage the
specifics of the firewall-driven package.  It revises the AOS Master's
class to incorporate this new module.  It also had to revise the extant
dhcpd-driven module and its integration so that both Django web apps
could be served from the same Apache httpd virtual host.

It might be better to leverage the concat module (from the Puppet Forge)
to assemble pieces of the httpd configuration with each of the web app
modules contributing their own specific parts, but this simpler approach
works well enough for now and does clearly illustrate the target setup.
To arrive at an intelligent setup using the concat module, I'll need
more time to evaluate the merits and issues that can only be had from
experience with a deployment.