Drop - dart::subsys::system_accounts
While this is a great concept, it failed in practice because its
impossible (AFAICS) to make such created user accounts members of groups
that only (and must) exist in LDAP.  In this case, the teamcity user
needed to become a member of the repomgr group.  The problems of dealing
with services which depend upon networked identities that may not be
available at the time of service startup, will have to be dealt with
more directly, such as through systemd ordering.