Change - use OpenJDK for TeamCity
I just realized that the default paths, alternatives and polar alignment
had it such that OpenJDK was being used here already, even though the
Oracle JDK had been intended.  Since the OpenJDK seems to be working so
well, that is now the intended JDK -- it's much easier to manage since
it's provided by Fedora.