Change - migrate TeamCity agent from Server to Agent3
Some build configurations (EST, NGIC or the like) require and expect
a PostgreSQL server.  Co-opting the one on the TeamCity server
(mdct-teamcity-f20) which was setup and intended only for TeamCity use
is unnecessarily risky.  To mitigate this risk, a new agent
(mdct-teamcity-agent3) has been established and the agent that ran on
the server has been disconnected, disabled and unauthorized.  Thus
agents now may do any database work necessary without worry of impacting
the TeamCity server's database.

Not shown here in this commit was a temporary "ensure => absent" of the
agent on the server.  That has already been applied and has no need for
remaining in the class.