Fix - cannot have multiple TeamCity installs running
For the TeamCity server, only one instance can bind the TCP port and
this ensures that it is the correct instance.

For the TeamCity agent, this ensures that the correct instance connects
to the server.  If the wrong instance was to connect instead, it will
block the correct instance (through licensing).  This also avoids any
odd mismatched behavior in crossing up versions.