New - puppet management of Bacula server
In the past, puppet only managed the Bacula server to the extent of
making it conform to a dart::abstract::guarded_server_node.  All the
actual setup of Bacula itself had been done manually (by Brendan
Schaenzer).  Since the Bacula server is being rebuilt at this time, I'm
taking the opportunity to rectify its construct here so that it's fully
managed.

This leverages the hundred-some hours of development for the bacula
module created at doubledog.org and does as little as possible to make
our old config fit into this module's structure.  It hasn't been tested
to be a viable replacement yet as I don't yet wish to involve the
storage array where the backup volumes are retained.  I first wish to
further improve the bacula module so that clients register their own
backup needs with the server.  This will improve the organization of
intent, improve accuracy and consistency and force recognition of loose
ends.