Drop - mdct_teamcity_f20 is no longer a build agent
In order to isolate the PostgreSQL installation on the TeamCity server
for TeamCity purposes from those required by build agents and their unit
tests which may expect a PostgreSQL server, the agent on the TeamCity
server has been previously de-registered within TeamCity.  This is
simply the other half that had since gone overlooked and was only
noticed because of a puppet error due to a duplicate declaration for the
PostgreSQL server.