Change - adjust Maximum Volumes downward
The intent had been to utilize more of the free disk space to achiever
longer retention when possible, but further analysis shows that the
number of volumes wasn't limited by this setting but rather the protocol
Bacula uses for recycling volumes.  It was able to achieve our minimum
retention periods with fewer volumes than our declared maximum.  While
Bacula doesn't recycle until required, it does seem to behave as if
there is a cost involved with creating more volumes (as there would be
for tape).  The apparent solution is to simply label additional volumes
manually so as to consume the free space, then when Bacula does require
recycling, the volumes will be that much older.

This downward revision is necessary after calculating what would be
possible given the total capacity of the file system, rather than the
approach last used which asked how many more volumes are possible given
the remaining free space.  Theoretically, 2143 volumes are possible, but
2100 was chosen here to allow a margin of safety.