Change - don't assume user's PyCharm edition
This creates a new module "dart::abstract::pycharm::professional" that
deploys the particular stable and EAP releases in use at Dart.  This
replaces the declarations within the "jetbrains::pycharm" module, which
now merely manages a host to use PyCharm but doesn't dictate any
editions or releases that are to be used.  In other words, Dart's needs
have been removed from the jetbrains::pycharm module.  (The same should
eventually happen for jetbrains::teamcity and jetbrains::idea.)