Fix - rpm-build-tools targets wrong repo
While this commit isn't any more correct, it's more usable.  Because
Dart has a dedicated "noarch" repository, it's inappropriate for
everything we typically package to land according to %ARCH.  Ideally
rpm-build-tools would sprout a new token that expands as does %ARCH, but
recognizes when a package is noarch in nature and targets accordingly.
Until then, this is an improvement.