New - slave support in "firewall_driven" module
Support for the firewall-driven-slave package has been added here and
the support for firewall-driven-master has been reworked for a better
combined fit.  Slave support is needed for AOS nodes, but the
integration has been done here since the module originated here; a copy
of the module will be pulled into the AOS conf.d resources later.