Fix - firewall_driven::slave::service ignores $enable
This parameter must be honored due to the stateless nature of the AOS.
AOS applications providing a systemd service unit lack an [Install]
section because it's important that such services be started only per
puppet.  Since they are "static" (systemd parlance for units lacking an
[Install] section), they cannot be disabled.

To achieve this, the service class had to be decoupled from the main
slave class (or relay even more parameters through and muddling their
sense).  Another alternative would have been to drop "enable" support
completely from the class but that have several downsides, among them:
    - It would have to be added back in and brought to this state again
      anyway should an [Install] section ever be added, such as might
      occur through use of Fedora Presets.
    - The class would be less useful as an ideal template model for
      non-AOS services.