Bug - NFS pipe file system is not managed
This was discovered during the initial setup of Garrett Schafsnitz's
workstation.  The NFS ID mapper (Class[nfs::rpcidmapd]) failed to start
because a number of mount points for RPC were missing under
/var/lib/nfs/rpc_pipefs.  These become available when the
var-lib-nfs-rpc_pipefs.mount systemd unit is active.  It is unclear why
this unit was inactive on his workstation, but rebooting clearly didn't
help.  Since this is critical to the operation of the ID mapper, it is
now being actively managed by puppet.