Refactor - simplify PICAPS Test host classes
There's absolutely no need for this class proliferation when all 20 host
classes are identical.  If one does need to be different, that would be
the time to implement such a class.  Otherwise, it's a much simpler
arrangement to have all PICAPS Test hosts be directly associated with
the abstract class.

This also replaces the 20 hostname-to-class associations with a single
association based on a strict regular-expression.

Lots of reduction, same great taste!